reorganizacion
Todos los archivos reorganizados, librería y bbdd a parte
